Initial commit of my generic Makefile
While there is still room for improvement, it is high time I got my
generic Makefile into source control. This started as a little
experiment on various approaches to typical Makefile problems, but it
has now started to grow into a generic solution that can be applied to
(simple) real life projects with only a little bit of customization.